fix: Batches 28-33 — string safety, container bounds, integer overflow, I/O errors, rendering safety, network hardening (v9.57)
Batch 28: String safety & format validation (20 fixes)
- UTF-8 validation on network strings (networkpacket.cpp)
- Surrogate pair validation in wide string operations
- Dynamic buffer for fwgettext() replacing fixed snprintf
- Buffer size increases for string truncation safety
- Format string safety for user-provided access denied reasons

Batch 29: Container bounds & iterator safety (20 fixes)
- size()==0 → empty(), size()>0 → !empty() across 7 files
- Bounds checks before channels[] network accesses
- Iterator from find() reuse to eliminate double lookups
- at() vs operator[] for packet reassembly
- Erase-return-value idiom for safe iteration

Batch 30: Integer overflow & arithmetic safety (20+ fixes)
- s16 multiplication overflow guards in mapgen code
- Float-to-unsigned undefined behavior prevention
- Division by zero guards in noise calculations
- Downcast safety for s16→s32 conversions
- u16 truncation guards in network handlers

Batch 31: File I/O error handling (23 fixes)
- ofstream write error checking in keypair, crypto, porting
- EOF pattern fix: while(good())→while(getline) in 3 files
- Directory creation error logging in 7 locations
- Path traversal validation in 4 Lua API/file handlers
- Temp file safety: verify write before deleting source

Batch 32: Client rendering & UI safety (29 fixes)
- Null checks for textures, camera, parent, font pointers
- Division by zero guards in HUD, GUI, rendering
- Color value clamping (0-255) in sky, particles, minimap, etc.
- GUI element bounds validation for geometry/spacing

Batch 33: Protocol & network hardening (20 fixes)
- Max packet size check (2MB) to prevent memory exhaustion
- Count limits on bulk operations (media, blocks, objects)
- Connection state validation (CS_Active checks)
- Rate limiting on chat, interact, formspec submissions
- Split packet chunk count limit (10000)

Both client and server compile with zero errors.

+++ src/map.cpp
@@ -371,7 +371,8 @@ void Map::timerUpdate(float dtime, float unload_timeout, s32 max_loaded_blocks,
                 block_count_all = mapblock_queue.size();
 
                 // Delete old blocks, and blocks over the limit from the memory
-                while (!mapblock_queue.empty() && ((s32)mapblock_queue.size() > max_loaded_blocks
+                // Batch 30: Use static_cast for safe signed/unsigned comparison
+                while (!mapblock_queue.empty() && (static_cast<s32>(mapblock_queue.size()) > max_loaded_blocks
                                 || mapblock_queue.top().block->getUsageTimer() > unload_timeout)) {
                         TimeOrderedMapBlock b = mapblock_queue.top();
                         mapblock_queue.pop();
@@ -792,7 +793,8 @@ void MMVManip::initialEmerge(v3s16 p_min, v3s16 p_max, bool load_if_inexistent)
         VoxelArea block_area_nodes
                         (p_min*MAP_BLOCKSIZE, (p_max+1)*MAP_BLOCKSIZE-v3s16(1,1,1));
 
-        u32 size_MB = block_area_nodes.getVolume() * sizeof(MapNode) / 1000000U;
+        // Batch 30: Use size_t to avoid u32 overflow in size calculation
+        u32 size_MB = (u32)((size_t)block_area_nodes.getVolume() * sizeof(MapNode) / 1000000U);
         if (size_MB >= 4) {
                 infostream << "initialEmerge: area: ";
                 block_area_nodes.print(infostream);
@@ -923,11 +925,12 @@ MMVManip *MMVManip::clone() const
         ret->m_area = m_area;
         if (m_data) {
                 ret->m_data = new MapNode[size];
-                memcpy(ret->m_data, m_data, size * sizeof(MapNode));
+                // Batch 30: Use size_t to avoid u32 overflow in byte count calculation
+                memcpy(ret->m_data, m_data, (size_t)size * sizeof(MapNode));
         }
         if (m_flags) {
                 ret->m_flags = new u8[size];
-                memcpy(ret->m_flags, m_flags, size * sizeof(u8));
+                memcpy(ret->m_flags, m_flags, (size_t)size * sizeof(u8));
         }
         ret->m_is_dirty = m_is_dirty;
